requirements.txt
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
attrs==25.4.0
bidict==0.23.1
bottle==0.13.4
certifi==2025.11.12
cffi==2.0.0
click==8.3.1
clr_loader==0.2.9
colorama==0.4.6
docutils==0.22.3
fastapi==0.124.4
frozenlist==1.8.0
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
ifaddr==0.2.0
itsdangerous==2.2.0
Jinja2==3.1.6
markdown2==2.5.4
MarkupSafe==3.0.3
mido==1.3.3
multidict==6.7.0
nicegui==3.4.0
orjson==3.11.5
packaging==25.0
platformdirs==4.5.1
propcache==0.4.1
proxy_tools==0.1.0
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
python-dotenv==1.2.1
python-engineio==4.12.3
python-multipart==0.0.20
python-rtmidi==1.5.8
python-socketio==5.15.0
pythonnet==3.0.5
pywebview==6.1
PyYAML==6.0.3
simple-websocket==1.1.0
starlette==0.50.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.38.0
watchfiles==1.1.1
websockets==15.0.1
wsproto==1.3.2
yarl==1.22.0
